requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
cx-Freeze==6.10
idna==3.4
importlib-metadata==5.1.0
packaging==21.3
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
urllib3==1.26.13
zipp==3.11.0
